requirements.txt
django
pillow
python-decouple
djangorestframework
markdown       
django-filter